move project to components
